[SHACK-304] Add Chef Apply as a gem dependency
ChefDK holds all the logic for the massive dependency resolution task.
Because of this we have to include Chef Apply as a gem dependency, even
though we only want to expose it via Chef Workstation and not Chef
Apply. That is why we do not appbundle it here but do appbundle it in
Chef Workstation.

As we move the dependency solving logic out of the ChefDK (per
https://github.com/chef/chef-rfc/pull/308) and seperate out components
(like policyfile and the 'chef' binary) we will eventually not need this
dependency here. But until that day...

Signed-off-by: tyler-ball <tball@chef.io>